requirements.txt
redis
numpy
scipy
matplotlib
pyyaml
pandas